# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing venders",
        "title": "Vender API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/invender": {
            "get": {
                "description": "Fetch all individual vender profiles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IndividualVender"
                ],
                "summary": "Get all individual venders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.IndvisulVenderProfile"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/invender/{vender_id}": {
            "get": {
                "description": "Fetch individual vender profile using vender_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IndividualVender"
                ],
                "summary": "Get individual vender by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vender ID",
                        "name": "vender_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IndvisulVenderProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update individual vender profile by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IndividualVender"
                ],
                "summary": "Update individual vender",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vender ID",
                        "name": "vender_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated vender data",
                        "name": "vender",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IndvisulVenderProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/invender/{vender_type}": {
            "post": {
                "description": "Create vender (INDIVISUAL / CLINIC / HOSPITAL)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IndividualVender"
                ],
                "summary": "Create vender by type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vender Type",
                        "name": "vender_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Individual vender payload",
                        "name": "vender",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IndvisulVenderProfile"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enum.VenderData": {
            "type": "string",
            "enum": [
                "UNVERIFIED",
                "VERIFIED",
                "REFIELD"
            ],
            "x-enum-varnames": [
                "UnVerified",
                "Verified",
                "ReField"
            ]
        },
...
        "model.IndvisulVenderProfile": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "phone_number",
                "vender_data"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "phone_number": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vender_data": {
                    "enum": [
                        "UNVERIFIED",
                        "VERIFIED",
                        "REFIELD"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.VenderData"
                        }
                    ]
                },
                "vender_id": {
                    "type": "string"
                }
            }
        },
...
    }
}